refactor(web): move shared helpers and provider marks out of settings
Missions, knowledge and pages imported errText, slugify, the provider
presets and ProviderLogo from the settings folder, which the design
contract reserves for settings-only code. The error helpers and slug
rule move to lib, provider presets to lib and ProviderLogo to the
canonical layer, so every feature imports shared shapes from a layer
none of them owns.
