+++ web/src/components/Message.tsx
@@ -13,7 +13,7 @@ import { ApprovalCard, type Decision } from './chat/ApprovalCard'
 import { ToolCallGroup } from './chat/ToolCallCard'
 import { AttachmentViewer, mimeLabel } from './AttachmentViewer'
 import { ModelBadge } from './ModelBadge'
-import { errText } from './settings/util'
+import { errText } from '../lib/errors'
 import { Badge } from './ui/badge'
 import { Button } from './ui/button'
 import { Alert, AlertTitle } from './ui/alert'
