+++ web/src/components/SessionList.tsx
@@ -14,8 +14,8 @@ import { Link, useLocation, useNavigate } from 'react-router'
 import { toast } from 'sonner'
 import { deleteSession, updateSession } from '../api/client'
 import type { SessionMeta } from '../api/types'
+import { errText } from '../lib/errors'
 import { groupByDay, useSessions } from '../lib/sessions'
-import { errText } from './settings/util'
 import { Badge } from './ui/badge'
 import { Button } from './ui/button'
 import {
