+++ web/src/components/missions/MissionForm.tsx
@@ -28,7 +28,7 @@ import type {
   Schedule,
 } from '../../api/types'
 import { useAgents, useRoutes } from '../AgentPicker'
-import { slugify } from '../settings/util'
+import { slugify } from '../../lib/slugify'
 import { cronPresets, type CronPresetValue, presetFor } from '../../lib/schedules'
 import { CURRENCIES } from '../../lib/currencies'
 import { extractRepoMentions, matchRepo } from '../../lib/goalRepo'
@@ -42,7 +42,7 @@ import { Label } from '../ui/label'
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../ui/select'
 import { Textarea } from '../ui/textarea'
-import { errText } from '../settings/util'
+import { errText } from '../../lib/errors'
 import { envIcon } from '../icons/EnvIcons'
 import { type PendingAttachment } from '../Composer'
 import { MissionAttachments } from './MissionAttachments'
