+++ web/src/components/missions/ReviewRoutePicker.tsx
@@ -3,7 +3,7 @@ import { toast } from 'sonner'
 import { getMissionExecutionPlan, patchMissionRouting } from '../../api/client'
 import type { ExecutionPlanEntry, Mission } from '../../api/types'
 import { useRoutes } from '../AgentPicker'
-import { errText } from '../settings/util'
+import { errText } from '../../lib/errors'
 import { Button } from '../ui/button'
 import { Input } from '../ui/input'
 import { Label } from '../ui/label'
