+++ web/src/components/settings/AgentForm.tsx
@@ -10,7 +10,8 @@ import {
 } from '../ui/select'
 import { Switch } from '../ui/switch'
 import { Field, FieldGroup } from '../timothy/field'
-import { slugify, UNSET } from './util'
+import { UNSET } from './util'
+import { slugify } from '../../lib/slugify'
 import { AllowlistPicker } from './AllowlistPicker'
 import { EXECUTOR_DEFAULT, executorChoices } from '../missions/MissionForm'
 import { listKbCollections, listSkills, listTools } from '../../api/client'
