+++ web/src/components/settings/AgentsList.tsx
@@ -14,7 +14,7 @@ import { PageHeader, SectionHeader } from '../timothy/page-header'
 import { PageShell } from '../timothy/page-shell'
 import { EntityCard } from './EntityCard'
 import { settingsArea } from './settingsAreas'
-import { errText } from './util'
+import { errText } from '../../lib/errors'
 
 const area = settingsArea('agents')
