+++ web/src/components/settings/ConnectorAdd.tsx
@@ -20,7 +20,9 @@ import { CredentialField, type CredentialMode } from './CredentialRefPicker'
 import { settingsArea } from './settingsAreas'
 import { TestStatus } from './TestStatus'
 import { useDefaultSecretBackend } from './useDefaultSecretBackend'
-import { connectedAs, errText, isTimothyAuthError, slugify } from './util'
+import { connectedAs } from './util'
+import { errText, isTimothyAuthError } from '../../lib/errors'
+import { slugify } from '../../lib/slugify'
 
 const area = settingsArea('connectors')
