+++ web/src/components/settings/FeaturesTab.tsx
@@ -14,7 +14,8 @@ import { Alert, AlertDescription } from '../ui/alert'
 import { PageHeader } from '../timothy/page-header'
 import { PageShell } from '../timothy/page-shell'
 import { settingsArea } from './settingsAreas'
-import { errText, UNSET } from './util'
+import { UNSET } from './util'
+import { errText } from '../../lib/errors'
 import { SettingFlagRow } from './SettingFlagRow'
 import { SettingValueCard } from './SettingValueCard'
