+++ web/src/components/settings/util.test.ts
@@ -1,33 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { humanizeProbeDetail, probeFailureText, responsesSuffix, slugify } from './util'
-
-describe('slugify', () => {
-  it('lowercases and hyphenates non-alphanumeric runs', () => {
-    expect(slugify('My Homelab Agent!')).toBe('my-homelab-agent')
-  })
-
-  it('trims leading and trailing hyphens', () => {
-    expect(slugify('  --Infra--  ')).toBe('infra')
-  })
-})
-
-describe('humanizeProbeDetail', () => {
-  it('extracts the message from an OpenAI-shaped JSON body', () => {
-    expect(
-      humanizeProbeDetail(
-        'http 401: {"error":{"code":"401","message":"token expired or incorrect"}}',
-      ),
-    ).toBe('Provider rejected the API key — token expired or incorrect')
-  })
-
-  it('leaves a non-JSON probe detail unchanged', () => {
-    expect(humanizeProbeDetail('upstream status 401')).toBe('upstream status 401')
-  })
-
-  it('uses the status label when the body is JSON without a message', () => {
-    expect(humanizeProbeDetail('http 429: {"error":{"code":"429"}}')).toBe('Rate limited')
-  })
-})
+import { probeFailureText, responsesSuffix } from './util'
 
 describe('probeFailureText', () => {
   it('does not render raw JSON in the connection-test banner', () => {
