+++ web/src/lib/slugify.ts
@@ -1,0 +1,7 @@
+// slugify mirrors the backend's name rule: lowercase slug.
+export function slugify(v: string): string {
+  return v
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, '-')
+    .replace(/^-+|-+$/g, '')
+}
