+++ web/src/pages/AutomationDetail.tsx
@@ -4,10 +4,10 @@ import { Link, useNavigate, useParams } from 'react-router'
 import { toast } from 'sonner'
 import { listDestinations, listMissions, listSchedules, patchSchedule } from '../api/client'
 import type { Destination, Mission, Schedule } from '../api/types'
-import { slugify } from '../components/settings/util'
 import { MissionCard } from '../components/missions/MissionCard'
 import { DestinationKindIcon } from '../components/destinations/DestinationKindIcon'
-import { errText } from '../components/settings/util'
+import { errText } from '../lib/errors'
+import { slugify } from '../lib/slugify'
 import { EmptyState } from '../components/timothy/empty-state'
 import { IconButton } from '../components/timothy/icon-button'
 import { PageHeader } from '../components/timothy/page-header'
